[rename] permissions, permission_doc -> validation, permissions
